fix: use DEFAULT_SET_ID constant and consolidate duplicate test file
- Replace literal "default" with DEFAULT_SET_ID in app.ts fallbacks
- Move MESSAGE_SNAPSHOT test from scenario-set-id-default into
  extensible-metadata suite, then delete the duplicate file
